# apps/api/rbac_dev-swagger-v1.0.0.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/svg/{name}": {
      "get": {
        "operationId": "AuthController_svgCaptcha_v1",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "验证名 [\"test\", \"register\", \"login\", \"resetPwd\", \"updateInfo\"]",
            "schema": {
              "example": "test",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SvgCaptchaVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "获取SVG验证码",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/email/{name}": {
      "post": {
        "operationId": "AuthController_emailCaptcha_v1",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "验证名 [\"test\", \"register\", \"login\", \"resetPwd\", \"updateInfo\"]",
            "schema": {
              "example": "test",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCaptchaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "发送邮件验证码",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/login/svg": {
      "post": {
        "operationId": "AuthController_svgLogin_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SvgLoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TokenVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "SVG登录",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/register/email": {
      "post": {
        "operationId": "AuthController_emailRegister_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRegisterDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "邮箱注册",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/login/email": {
      "post": {
        "operationId": "AuthController_emailLogin_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TokenVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "邮箱登录",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/reset-pwd/email": {
      "post": {
        "operationId": "AuthController_emailResetPwd_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailResetPwdDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "邮箱重置密码",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TokenVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "刷新令牌",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_loginOut_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "登出",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_meInfo_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserDetailsVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "获取当前登录用户信息",
        "tags": ["Auth"]
      }
    },
    "/v1/auth": {
      "post": {
        "operationId": "AuthController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthDetailsVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "创建认证",
        "tags": ["Auth"]
      },
      "get": {
        "operationId": "AuthController_list_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "第几页",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "一页几条数据",
            "schema": {
              "minimum": 1,
              "maximum": 2000,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "搜索关键词",
            "schema": {
              "example": "张三",
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "排序字段(创建时间:createdAt 更新时间:updatedAt 名称:name)",
            "schema": {
              "example": "createdAt",
              "enum": ["createdAt", "updatedAt", "name"],
              "type": "string"
            }
          },
          {
            "name": "orderType",
            "required": false,
            "in": "query",
            "description": "排序方式(升序:asc 降序:desc)",
            "schema": {
              "example": "asc",
              "enum": ["asc", "desc"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FindAllAuthVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "获取认证列表",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/{id}": {
      "delete": {
        "operationId": "AuthController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "业务ID",
            "schema": {
              "minLength": 36,
              "maxLength": 36,
              "example": "xxx",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "删除认证",
        "tags": ["Auth"]
      },
      "patch": {
        "operationId": "AuthController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "业务ID",
            "schema": {
              "minLength": 36,
              "maxLength": 36,
              "example": "xxx",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "更新认证",
        "tags": ["Auth"]
      },
      "get": {
        "operationId": "AuthController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "业务ID",
            "schema": {
              "minLength": 36,
              "maxLength": 36,
              "example": "xxx",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthDetailsVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "查看单个认证详情",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/{id}/status": {
      "patch": {
        "operationId": "AuthController_updateStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "业务ID",
            "schema": {
              "minLength": 36,
              "maxLength": 36,
              "example": "xxx",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "更新认证状态",
        "tags": ["Auth"]
      }
    },
    "/v1/auth/{id}/sort": {
      "patch": {
        "operationId": "AuthController_updateSort_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "业务ID",
            "schema": {
              "minLength": 36,
              "maxLength": 36,
              "example": "xxx",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSortDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResVO"
                    },
                    {
                      "properties": {
                        "data": {
                          "example": []
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          },
          {
            "COOKIE-JWT": []
          }
        ],
        "summary": "更新认证排序优先级",
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "rbac_dev",
    "description": "rbac_dev 项目的接口文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "在请求头Authorization中携带JWT，格式: Bearer <JWT_TOKEN>"
      },
      "COOKIE-JWT": {
        "type": "apiKey",
        "in": "cookie",
        "name": "Authorization",
        "description": "在Cookie Authorization中携带JWT，格式: Bearer <JWT_TOKEN>"
      }
    },
    "schemas": {
...
      "SvgCaptchaVO": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "svg验证码凭证",
            "example": "svg验证码凭证"
          },
          "svg": {
            "type": "string",
            "description": "svg验证码Base64",
            "example": "data:image/svg+xml;base64,XXXX"
          }
        },
        "description": "svg验证码",
        "required": ["token", "svg"]
      },
      "TokenVO": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "访问令牌"
          },
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌"
          }
        },
        "description": "令牌详情",
        "required": ["accessToken"]
      },
      "EmailCaptchaDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "zhangsan@example.com",
            "minLength": 2,
            "maxLength": 191
          }
        },
        "description": "邮箱验证码接口参数校验",
        "required": ["email"]
      },
      "SvgLoginDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "用户名",
            "example": "zhangsan",
            "minLength": 2,
            "maxLength": 64
          },
          "pwd": {
            "type": "string",
            "description": "密码",
            "example": "Aa123456",
            "minLength": 8,
            "maxLength": 64
          },
          "token": {
            "type": "string",
            "description": "验证码令牌",
            "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
            "minLength": 36,
            "maxLength": 36
          },
          "captcha": {
            "type": "string",
            "description": "验证码",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": ["name", "pwd", "token", "captcha"]
      },
      "EmailRegisterDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "用户名",
            "example": "zhangsan",
            "minLength": 2,
            "maxLength": 64
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "zhangsan@example.com",
            "minLength": 2,
            "maxLength": 191
          },
          "pwd": {
            "type": "string",
            "description": "密码",
            "example": "Aa123456",
            "minLength": 8,
            "maxLength": 64
          },
          "captcha": {
            "type": "string",
            "description": "验证码",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "description": "邮箱注册接口参数校验",
        "required": ["name", "email", "pwd", "captcha"]
      },
      "EmailLoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "zhangsan@example.com",
            "minLength": 2,
            "maxLength": 191
          },
          "pwd": {
            "type": "string",
            "description": "密码",
            "example": "Aa123456",
            "minLength": 8,
            "maxLength": 64
          },
          "captcha": {
            "type": "string",
            "description": "验证码",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "description": "邮箱登录接口参数校验",
        "required": ["email", "pwd", "captcha"]
      },
      "EmailResetPwdDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "zhangsan@example.com",
            "minLength": 2,
            "maxLength": 191
          },
          "pwd": {
            "type": "string",
            "description": "密码",
            "example": "Aa123456",
            "minLength": 8,
            "maxLength": 64
          },
          "captcha": {
            "type": "string",
            "description": "验证码",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          },
          "confirmPwd": {
            "type": "string",
            "description": "确认密码",
            "example": "Aa123456",
            "minLength": 8,
            "maxLength": 64
          }
        },
        "description": "邮箱重置密码接口参数校验",
        "required": ["email", "pwd", "captcha", "confirmPwd"]
      },
      "RefreshTokenDTO": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌",
            "example": "xxx",
            "minLength": 271,
            "maxLength": 271
          }
        },
        "description": "刷新令牌接口参数校验"
      },
      "CreateAuthDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "认证名",
            "example": "认证名",
            "minLength": 2,
            "maxLength": 64
          },
          "remark": {
            "type": "string",
            "description": "备注",
            "example": "xxx",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "description": "认证名参数校验",
        "required": ["name"]
      },
      "UpdateAuthDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "认证名",
            "example": "认证名",
            "minLength": 2,
            "maxLength": 64
          },
          "remark": {
            "type": "string",
            "description": "备注",
            "example": "xxx",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "description": "更新认证接口参数校验"
      },
      "ResVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "业务码",
            "example": "0"
          },
          "msg": {
            "type": "string",
            "description": "业务信息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "业务数据(对象或数组)"
          },
          "originUrl": {
            "type": "string",
            "description": "请求地址",
            "example": "xxx"
          },
          "referer": {
            "type": "string",
            "description": "请求源",
            "example": "xxx"
          },
          "userAgent": {
            "type": "string",
            "description": "客户端信息",
            "example": "xxx"
          },
          "timestamp": {
            "type": "number",
            "description": "时间戳",
            "example": 1672531200000
          },
          "clientIp": {
            "type": "string",
            "description": "客户端IP",
            "example": "xxx"
          }
        },
        "description": "响应格式化",
        "required": ["code", "msg", "data", "originUrl", "referer", "userAgent", "timestamp", "clientIp"]
      }
    }
  }
}